Move allocation state out of statusmanager
